svc_encoder_rtc.cc: fix conversion warning
w/various versions of Visual Studio, after:
5665f1cd6d svc_encoder_rtc: improve multilayer metadata handling

fixes:
examples\svc_encoder_rtc.cc(1416,62): warning C4244:
  'argument': conversion from 'const uint16_t' to 'uint8_t', possible
  loss of data

+++ examples/svc_encoder_rtc.cc
@@ -1413,7 +1413,8 @@ static void write_depth_representation_element(
     die("Invalid mantissan_len %d\n", element.first.mantissa_len);
   }
   write_literal(buffer, element.first.mantissa_len - 1, 5);
-  write_literal(buffer, element.first.mantissa, element.first.mantissa_len);
+  write_literal(buffer, element.first.mantissa,
+                static_cast<uint8_t>(element.first.mantissa_len));
 }
 
 static void write_color_properties(
